+++ src/main/java/com/vnano/defenders/enchantment/ModEnchantments.java
@@ -1,5 +1,8 @@
 package com.vnano.defenders.enchantment;
 
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.item.ItemStack;
+
 public final class ModEnchantments {
     public static EnchantmentDefender FOOTWORK;
     public static EnchantmentDefender FORTIFICATION;
@@ -7,6 +10,11 @@ public final class ModEnchantments {
     public static EnchantmentDefender FINESSE;
     public static EnchantmentDefender REFLEXES;
     public static EnchantmentDefender DEFLECTION;
+
+    public static int getLevel(EnchantmentDefender enchantment, ItemStack stack) {
+        return enchantment != null && enchantment.isEnabled()
+            ? EnchantmentHelper.getEnchantmentLevel(enchantment, stack) : 0;
+    }
 
     private ModEnchantments() {}
 }